fix ci errors
