+++ tests/cpp/device_config.cpp
@@ -20,12 +20,12 @@ TEST(TestDeviceConfig, kv_cache_precision_u8) {
     size_t head_size = 64, head_size_u8 = head_size + 8;
     std::vector<size_t> num_kv_heads(12, 12);
 
-    ov::genai::DeviceConfig device_config_default(core, scheduler_config, "CPU");
-    device_config_default.set_model_params(num_kv_heads, head_size_u8, num_decoder_layers);
+    ov::genai::DeviceConfig device_config_f16(core, scheduler_config, "CPU", { ov::hint::kv_cache_precision(ov::element::f16) });
+    device_config_f16.set_model_params(num_kv_heads, head_size_u8, num_decoder_layers);
 
-    ov::genai::DeviceConfig device_config_u8(core, scheduler_config, "CPU", { ov::hint::kv_cache_precision(ov::element::u8) });
+    ov::genai::DeviceConfig device_config_u8(core, scheduler_config, "CPU");
     device_config_u8.set_model_params(num_kv_heads, head_size, num_decoder_layers);
 
     const auto ratio = ov::element::f16.size() / ov::element::u8.size();
-    ASSERT_EQ(device_config_default.get_num_kv_blocks() * ratio, device_config_u8.get_num_kv_blocks());
+    ASSERT_EQ(device_config_f16.get_num_kv_blocks() * ratio, device_config_u8.get_num_kv_blocks());
 }
